+++ app/entrypoints/install_dustmaps_config.sh
@@ -1,0 +1,5 @@
+# If .dustmapsrc has been mounted as a read-only file at /tmp/.dustmapsrc,
+# copy it to the expected location /root/.dustmapsrc if it does not already exist.
+if [[ ! -f "/root/.dustmapsrc" && -f "/tmp/.dustmapsrc" ]]; then
+  cp /tmp/.dustmapsrc /root/.dustmapsrc
+fi